pyproject.toml
[tool.poetry]
name = "pytorch-image"
version = "1.0.0"
description = "Minimalistic library designed to be an interface for the model input with a lightweight `Image` class."
authors = ["Richard Löwenström <samedii@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8,<3.11"
torch = "*"
numpy = "*"
pillow = "^10.4.0"
pydantic = "^2.9.2"

[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
ipywidgets = "^8.1.5"
pytest = "^8.3.3"
ruff = "^0.6.8"

[tool.poetry.group.train.dependencies]
albumentations = "^1.4.16"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
